Adds column to users and relationship to players
Adds a column to the users resource called accnt_type.  This will be
used to determine if the user is a coach or admin.  Admins will be able
to create seasons in future versions of the app.

Adds the relationship users have many players necessary for MVP

Adds an an accnt_type parameter to the sign-up curl script