Adds curl scripts and too much other code
Adds example curl scripts for player and season.
Adds controller, model, and serializer files for season resource.
  these were not created with scaffold.  didn't notice until late so
  they were created manually
Updates the serializer for player to fix problem where I was only seeing
the player ID in the browser

Adds changes in players_controller and seasons_controller where there
was a missing 'end' statement.

Adds a batting order field to players (discovered that this was not part
of the initial db creation during curl scipt testing)

Adds user foreign key to seasons